:recycle: Split transaction pool and change logic to be more functional
